Project looms, v.0000001a
This is the first commit of the looms project.  It has very basic, very
primitive, and largely very poor functionality.  Current system is entirely
coupled to a single database system (MySQL) with the worlds most awful
schema, and is entirely tied to Ansibl as a command transport layer.

There are all kinds of changes that will be created in the near future,
such as decoupling the tools from one single database, adding support
for more than just Debian package repositories, etc etc etc.  But, for now,
this is what we have to work with.